chore: seperate classes to their own files
